+++ eras/conway/impl/src/Cardano/Ledger/Conway/TxInfo.hs
@@ -33,6 +33,7 @@ module Cardano.Ledger.Conway.TxInfo (
   transValidityInterval,
   transVotingProcedures,
   transProposal,
+  transVoter,
   toPlutusV3Args,
   transTxCertV1V2,
   transPlutusPurposeV1V2,
@@ -48,6 +49,7 @@ import Cardano.Ledger.Alonzo.Plutus.Context (
   EraPlutusContext (..),
   EraPlutusTxInfo (..),
   LedgerTxInfo (..),
+  PlutusPurposeScriptHashArg,
   PlutusTxCert,
   PlutusTxInfoResult (..),
   SupportedLanguage (..),
@@ -60,7 +62,10 @@ import Cardano.Ledger.Alonzo.Plutus.TxInfo (
  )
 import qualified Cardano.Ledger.Alonzo.Plutus.TxInfo as Alonzo
 import Cardano.Ledger.Alonzo.Scripts (AlonzoPlutusPurpose (..), toAsItem)
-import Cardano.Ledger.Babbage.TxInfo (BabbageContextError (..), transTxOutV2)
+import Cardano.Ledger.Babbage.TxInfo (
+  BabbageContextError (..),
+  transTxOutV2,
+ )
 import qualified Cardano.Ledger.Babbage.TxInfo as Babbage
 import Cardano.Ledger.BaseTypes (
   Inject (..),
@@ -445,6 +450,10 @@ instance EraPlutusTxInfo 'PlutusV1 ConwayEra where
 
   toPlutusTxInInfo _ = transTxInInfoV1
 
+  toPlutusRedeemerPointer = Alonzo.transRedeemerPointerV1
+
+  toPlutusTxOut _ src txOut = Just <$> transTxOutV1 src txOut
+
 instance EraPlutusTxInfo 'PlutusV2 ConwayEra where
   toPlutusTxCert _ _ = transTxCertV1V2
 
@@ -464,7 +473,7 @@ instance EraPlutusTxInfo 'PlutusV2 ConwayEra where
           [minBound ..]
           (F.toList (txBody ^. outputsTxBodyL))
       txCerts <- Alonzo.transTxBodyCerts proxy ltiProtVer txBody
-      plutusRedeemers <- Babbage.transTxRedeemers proxy ltiProtVer tx
+      plutusRedeemers <- Babbage.transTxRedeemers proxy ltiProtVer tx ltiUTxO
       -- It is important for memoization for `txInfo` to be a let binding
       let
         txInfo =
@@ -488,6 +497,10 @@ instance EraPlutusTxInfo 'PlutusV2 ConwayEra where
 
   toPlutusTxInInfo _ = Babbage.transTxInInfoV2
 
+  toPlutusRedeemerPointer = Babbage.transRedeemerPointerV2V3
+
+  toPlutusTxOut _ = transTxOutV2
+
 instance EraPlutusTxInfo 'PlutusV3 ConwayEra where
   toPlutusTxCert _ pv = pure . transTxCert pv
 
@@ -510,7 +523,7 @@ instance EraPlutusTxInfo 'PlutusV3 ConwayEra where
           [minBound ..]
           (F.toList (txBody ^. outputsTxBodyL))
       txCerts <- Alonzo.transTxBodyCerts proxy ltiProtVer txBody
-      plutusRedeemers <- Babbage.transTxRedeemers proxy ltiProtVer tx
+      plutusRedeemers <- Babbage.transTxRedeemers proxy ltiProtVer tx ltiUTxO
       -- It is important for memoization for `txInfo` to be a let binding
       let txInfo =
             PV3.TxInfo
@@ -541,6 +554,10 @@ instance EraPlutusTxInfo 'PlutusV3 ConwayEra where
   toPlutusArgs = toPlutusV3Args
 
   toPlutusTxInInfo _ = transTxInInfoV3
+
+  toPlutusRedeemerPointer = Babbage.transRedeemerPointerV2V3
+
+  toPlutusTxOut _ = transTxOutV2
 
 transTxId :: TxId -> PV3.TxId
 transTxId txId = PV3.TxId (transSafeHash (unTxId txId))
@@ -640,9 +657,10 @@ transPlutusPurposeV3 ::
   ) =>
   proxy 'PlutusV3 ->
   ProtVer ->
+  () ->
   PlutusPurpose AsIxItem era ->
   Either (ContextError era) PV3.ScriptPurpose
-transPlutusPurposeV3 proxy pv = \case
+transPlutusPurposeV3 proxy pv _ = \case
   SpendingPurpose (AsIxItem _ txIn) -> pure $ PV3.Spending (transTxIn txIn)
   MintingPurpose (AsIxItem _ policyId) -> pure $ PV3.Minting (Alonzo.transPolicyID policyId)
   CertifyingPurpose (AsIxItem ix txCert) ->
@@ -733,18 +751,20 @@ transProposal proxy ProposalProcedure {pProcDeposit, pProcReturnAddr, pProcGovAc
 transPlutusPurposeV1V2 ::
   forall l era proxy.
   ( PlutusTxCert l ~ PV2.DCert
+  , PlutusPurposeScriptHashArg l ~ ()
   , EraPlutusTxInfo l era
   , Inject (ConwayContextError era) (ContextError era)
   ) =>
   proxy l ->
   ProtVer ->
+  () ->
   PlutusPurpose AsIxItem era ->
   Either (ContextError era) PV2.ScriptPurpose
-transPlutusPurposeV1V2 proxy pv = \case
-  SpendingPurpose asIxItem -> Alonzo.transPlutusPurpose proxy pv $ AlonzoSpending asIxItem
-  MintingPurpose asIxItem -> Alonzo.transPlutusPurpose proxy pv $ AlonzoMinting asIxItem
-  CertifyingPurpose asIxItem -> Alonzo.transPlutusPurpose proxy pv $ AlonzoCertifying asIxItem
-  WithdrawingPurpose asIxItem -> Alonzo.transPlutusPurpose proxy pv $ AlonzoWithdrawing asIxItem
+transPlutusPurposeV1V2 proxy pv sh = \case
+  SpendingPurpose asIxItem -> Alonzo.transPlutusPurpose proxy pv sh $ AlonzoSpending asIxItem
+  MintingPurpose asIxItem -> Alonzo.transPlutusPurpose proxy pv sh $ AlonzoMinting asIxItem
+  CertifyingPurpose asIxItem -> Alonzo.transPlutusPurpose proxy pv sh $ AlonzoCertifying asIxItem
+  WithdrawingPurpose asIxItem -> Alonzo.transPlutusPurpose proxy pv sh $ AlonzoWithdrawing asIxItem
   purpose -> Left $ inject $ PlutusPurposeNotSupported @era $ hoistPlutusPurpose toAsItem purpose
 
 transProtVer :: ProtVer -> PV3.ProtocolVersion
@@ -755,13 +775,14 @@ toPlutusV3Args ::
   EraPlutusTxInfo 'PlutusV3 era =>
   proxy 'PlutusV3 ->
   ProtVer ->
+  ScriptHash ->
   PV3.TxInfo ->
   PlutusPurpose AsIxItem era ->
   Maybe (Data era) ->
   Data era ->
   Either (ContextError era) (PlutusArgs 'PlutusV3)
-toPlutusV3Args proxy pv txInfo plutusPurpose maybeSpendingData redeemerData = do
-  scriptPurpose <- toPlutusScriptPurpose proxy pv plutusPurpose
+toPlutusV3Args proxy pv _ txInfo plutusPurpose maybeSpendingData redeemerData = do
+  scriptPurpose <- toPlutusScriptPurpose proxy pv () plutusPurpose
   let scriptInfo =
         scriptPurposeToScriptInfo
           scriptPurpose