+++ libs/plutus-preprocessor/src/Cardano/Ledger/Plutus/Preprocessor.hs
@@ -14,6 +14,7 @@ import Cardano.Ledger.Plutus.Language (
 import qualified Cardano.Ledger.Plutus.Preprocessor.Binary.V1 as V1
 import qualified Cardano.Ledger.Plutus.Preprocessor.Binary.V2 as V2
 import qualified Cardano.Ledger.Plutus.Preprocessor.Binary.V3 as V3
+import qualified Cardano.Ledger.Plutus.Preprocessor.Binary.V4 as V4
 import Data.ByteString.Short as SBS (fromShort)
 import Data.Foldable (forM_)
 import Data.List (intercalate)
@@ -99,7 +100,7 @@ allTestScripts =
         PlutusV1 -> Just V1.alwaysSucceedsNoDatumBytes
         PlutusV2 -> Just V2.alwaysSucceedsNoDatumBytes
         PlutusV3 -> Just V3.alwaysSucceedsNoDatumBytes
-        PlutusV4 -> Just V3.alwaysSucceedsNoDatumBytes
+        PlutusV4 -> Just V4.alwaysSucceedsNoDatumBytes
     , pure "Script that always succeeds, unless arguments are malformed or context contains a datum"
     )
   ,
@@ -108,7 +109,7 @@ allTestScripts =
         PlutusV1 -> Just V1.alwaysSucceedsWithDatumBytes
         PlutusV2 -> Just V2.alwaysSucceedsWithDatumBytes
         PlutusV3 -> Just V3.alwaysSucceedsWithDatumBytes
-        PlutusV4 -> Just V3.alwaysSucceedsWithDatumBytes
+        PlutusV4 -> Just V4.alwaysSucceedsWithDatumBytes
     , pure
         "Script that always succeeds, unless arguments are malformed or context does not contain a datum"
     )
@@ -118,7 +119,7 @@ allTestScripts =
         PlutusV1 -> Just V1.alwaysFailsNoDatumBytes
         PlutusV2 -> Just V2.alwaysFailsNoDatumBytes
         PlutusV3 -> Just V3.alwaysFailsNoDatumBytes
-        PlutusV4 -> Just V3.alwaysFailsNoDatumBytes
+        PlutusV4 -> Just V4.alwaysFailsNoDatumBytes
     , pure "Script that always fails, unless arguments are malformed or context contains a datum"
     )
   ,
@@ -127,7 +128,7 @@ allTestScripts =
         PlutusV1 -> Just V1.alwaysFailsWithDatumBytes
         PlutusV2 -> Just V2.alwaysFailsWithDatumBytes
         PlutusV3 -> Just V3.alwaysFailsWithDatumBytes
-        PlutusV4 -> Just V3.alwaysFailsWithDatumBytes
+        PlutusV4 -> Just V4.alwaysFailsWithDatumBytes
     , pure "Script that always fails, unless arguments are malformed or context does not contain a datum"
     )
   ,
@@ -136,7 +137,7 @@ allTestScripts =
         PlutusV1 -> Just V1.redeemerSameAsDatumBytes
         PlutusV2 -> Just V2.redeemerSameAsDatumBytes
         PlutusV3 -> Just V3.redeemerSameAsDatumBytes
-        PlutusV4 -> Just V3.redeemerSameAsDatumBytes
+        PlutusV4 -> Just V4.redeemerSameAsDatumBytes
     , "Script that succeeds whenever redeemer equals to the datum"
         :| [ "Fails on malformed arguments"
            ]
@@ -147,7 +148,7 @@ allTestScripts =
         PlutusV1 -> Just V1.evenDatumBytes
         PlutusV2 -> Just V2.evenDatumBytes
         PlutusV3 -> Just V3.evenDatumBytes
-        PlutusV4 -> Just V3.evenDatumBytes
+        PlutusV4 -> Just V4.evenDatumBytes
     , "Script that succeeds whenever Integer datum is supplied and it's value is even."
         :| [ "Fails on malformed arguments"
            ]
@@ -158,7 +159,7 @@ allTestScripts =
         PlutusV1 -> Just V1.evenRedeemerNoDatumBytes
         PlutusV2 -> Just V2.evenRedeemerNoDatumBytes
         PlutusV3 -> Just V3.evenRedeemerNoDatumBytes
-        PlutusV4 -> Just V3.evenRedeemerNoDatumBytes
+        PlutusV4 -> Just V4.evenRedeemerNoDatumBytes
     , "Script that succeeds whenever Integer redeemer is supplied and it's value is even"
         :| [ "Fails on malformed arguments or whenever datum is present in the context"
            ]
@@ -169,7 +170,7 @@ allTestScripts =
         PlutusV1 -> Just V1.evenRedeemerWithDatumBytes
         PlutusV2 -> Just V2.evenRedeemerWithDatumBytes
         PlutusV3 -> Just V3.evenRedeemerWithDatumBytes
-        PlutusV4 -> Just V3.evenRedeemerWithDatumBytes
+        PlutusV4 -> Just V4.evenRedeemerWithDatumBytes
     , "Script that succeeds whenever Integer redeemer is supplied and it's value is even"
         :| [ "Fails on malformed arguments or whenever datum is missing from the context"
            ]
@@ -180,7 +181,7 @@ allTestScripts =
         PlutusV1 -> Just V1.purposeIsWellformedNoDatumBytes
         PlutusV2 -> Just V2.purposeIsWellformedNoDatumBytes
         PlutusV3 -> Just V3.purposeIsWellformedNoDatumBytes
-        PlutusV4 -> Just V3.purposeIsWellformedNoDatumBytes
+        PlutusV4 -> Just V4.purposeIsWellformedNoDatumBytes
     , "Script that succeeds when datum is not expected and purpose arguments are validated against txInfo"
         :| [ "Fails on malformed arguments"
            ]
@@ -191,7 +192,7 @@ allTestScripts =
         PlutusV1 -> Just V1.purposeIsWellformedWithDatumBytes
         PlutusV2 -> Just V2.purposeIsWellformedWithDatumBytes
         PlutusV3 -> Just V3.purposeIsWellformedWithDatumBytes
-        PlutusV4 -> Just V3.purposeIsWellformedWithDatumBytes
+        PlutusV4 -> Just V4.purposeIsWellformedWithDatumBytes
     , "Script that succeeds when datum is expected and purpose arguments are validated against txInfo"
         :| [ "Fails on malformed arguments"
            ]
@@ -202,7 +203,7 @@ allTestScripts =
         PlutusV1 -> Just V1.datumIsWellformedBytes
         PlutusV2 -> Just V2.datumIsWellformedBytes
         PlutusV3 -> Just V3.datumIsWellformedBytes
-        PlutusV4 -> Just V3.datumIsWellformedBytes
+        PlutusV4 -> Just V4.datumIsWellformedBytes
     , "Script that succeeds when datum is expected and datum is validated against txInfo"
         :| [ "Fails on malformed arguments"
            ]
@@ -213,7 +214,7 @@ allTestScripts =
         PlutusV1 -> Just V1.inputsOutputsAreNotEmptyNoDatumBytes
         PlutusV2 -> Just V2.inputsOutputsAreNotEmptyNoDatumBytes
         PlutusV3 -> Just V3.inputsOutputsAreNotEmptyNoDatumBytes
-        PlutusV4 -> Just V3.inputsOutputsAreNotEmptyNoDatumBytes
+        PlutusV4 -> Just V4.inputsOutputsAreNotEmptyNoDatumBytes
     , "Script that succeeds when inputs and outputs are not empty validated against txInfo"
         :| [ "Fails on malformed arguments and also if inputs or outputs are empty"
            ]
@@ -224,7 +225,7 @@ allTestScripts =
         PlutusV1 -> Just V1.inputsOutputsAreNotEmptyWithDatumBytes
         PlutusV2 -> Just V2.inputsOutputsAreNotEmptyWithDatumBytes
         PlutusV3 -> Just V3.inputsOutputsAreNotEmptyWithDatumBytes
-        PlutusV4 -> Just V3.inputsOutputsAreNotEmptyWithDatumBytes
+        PlutusV4 -> Just V4.inputsOutputsAreNotEmptyWithDatumBytes
     , "Script that succeeds when inputs and outputs are not empty validated against txInfo"
         :| [ "Fails on malformed arguments and also if inputs or outputs are empty"
            ]
@@ -235,7 +236,7 @@ allTestScripts =
         PlutusV1 -> Nothing
         PlutusV2 -> Just V2.inputsOverlapsWithRefInputsBytes
         PlutusV3 -> Just V3.inputsOverlapsWithRefInputsBytes
-        PlutusV4 -> Just V3.inputsOverlapsWithRefInputsBytes
+        PlutusV4 -> Just V4.inputsOverlapsWithRefInputsBytes
     , "Script that succeeds only if any the inputs also appears in the reference inputs" :| []
     )
   ]