+++ eras/dijkstra/impl/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.4.0.0
 
+* Change `PointerPresentInOutput` constructor of `DijkstraContextError` to contain a `NonEmptySet TxOutSource` instead of `NonEmpty (TxOut era)`
+* Add `udppPlutusV4CostModel` field to `UpgradeDijkstraPParams`
+* Add `HKDApplicative` constraint to `upgradeDijkstraPParams`
+* Add `transRedeemerPointerV4`
 * Add `EncCBOR`, `ToCBOR` for `Block`
 * Add `DecCBOR` instances for `Annotator Block`
 * Remove `EncCBORGroup` instance for `DijkstraBlockBody`