+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -11,6 +11,9 @@
   - `checkFailOnNonEmpty` -> `failOnNonEmpty`
   - `checkFailOnNonEmptySet` -> `failOnNonEmptySet`
   - `checkFailOnNonEmptyMap` -> `failOnNonEmptyMap`
+* Add `Ord` instance for `TxOutSource`
+* Change `PlutusArgs 'PlutusV4` to contain `PV4.ScriptContext` instead of `PV3.ScriptContext`
+* Add `NFData` instance for `PV4.ScriptContext`
 * Remove `numSegComponents` from `EraBlockBody`
 * Remove generic `EncCBOR`, `ToCBOR`, and `DecCBOR` instances for `Block` in favor of per-era instances
 * Change `decodeMetadatum` to require definite-length chunks in indefinite-length bytestrings